Refactor: move SessionProvider to contexts folder
- Move session-provider.tsx from components to contexts/session-context.tsx
- Rename useSession hook to useAppSession to avoid conflict with next-auth
- Fix null check bug (was checking undefined, context default is null)
- Remove console.log debug statement
- Update barrel exports in contexts/index.ts
- Update imports in auth-wrapper.tsx and header.tsx
