+++ src/contexts/index.ts
@@ -1,1 +1,2 @@
 export { UserProvider, useUser } from "./user-context";
+export { SessionProvider, useAppSession } from "./session-context";
